requirements.txt
requests==2.28.2
pytest==7.3.0
allure-pytest==2.13.1
